move `position-data.ts` to `lib/trade-executor/state/`
