+++ src/lib/trade-executor/state/position-data.ts
